Attempt at fixing it for Sierra
I tried the original code on a Sierra machine and was able to open the menu.

Unfortunately, I didn't have any AppleTV on the network to try it out fully but
the basic AppleScript seems valid. (I tried with "Open Displays Preferences…")

Opening that up to see if it can solve #3.