Actual fix for Sierra
It seems that the menu options don't update right away when opening the
menu. Not sure why...
In any case, I added a delay and dealt with finding the item
differently.
Again, no idea why.
It seems to work.

...mostly. For some reason, when trying it out, the auto-completion
would tend to suggest "Apple TV (2)", but the menu would sometimes show
"Apple TV" and sometimes "Apple TV (2)". When that happens, the menu
stays open. Definitely not great but better than not working at all for
now.